contrib/ec-metric-gen/requirements.txt
Flask
requests
pyyaml
timeeval-gutenTAG
prometheus_client
flask_swagger_ui
